+++ web/app/react/components/Application.js
@@ -1,18 +1,61 @@
+import PropTypes from 'prop-types';
 import React from 'react';
 import {connect} from 'react-redux';
+import styled from 'styled-components';
 
+import {toggleChrome} from 'app/stores/application/applicationActions';
+
+import FormCheckableInput from 'src/components/Form/FormCheckableInput';
 import Tabs from 'src/components/Tabs/Tabs';
 import Tab from 'src/components/Tabs/Tab';
 
 import MainNavigation from './MainNavigation';
 import MainNavigationSection from './MainNavigationSection';
 
+export const ChromeToggleContainerStyled = styled.div`
+  position: fixed;
+  bottom: 10px;
+  right: 10px;
+  opacity: 0.1;
+  transition: opacity 0.15s linear;
+
+  &:hover {
+    opacity: 1;
+  }
+`;
+
+export const mapStateToProps = function({menu, application}) {
+  return {
+    menuItems: menu.items,
+    showChrome: application.showChrome,
+  };
+};
+
+export const mapDispatchToProps = {
+  toggleChrome,
+};
+
 class Application extends React.Component {
+  static propTypes = {
+    menuItems: PropTypes.array.isRequired,
+    showChrome: PropTypes.bool.isRequired,
+    toggleChrome: PropTypes.func.isRequired,
+  };
+
   onClickGitHub = () => {
     window.open('https://github.com/ryanzec/underpin-ui', '_blank');
   };
 
+  onToggleChrome = () => {
+    console.log('test');
+    this.props.toggleChrome();
+  };
+
   renderHeader() {
+    if (!this.props.showChrome) {
+      return null;
+    }
+
     return (
       <div className="application-container__header">
         <div className="application-container__header-logo">Underpin UI</div>
@@ -41,6 +84,10 @@ class Application extends React.Component {
   }
 
   renderMenu() {
+    if (!this.props.showChrome) {
+      return null;
+    }
+
     let menuNode = null;
     const menuSectionNodes = [];
 
@@ -69,15 +116,18 @@ class Application extends React.Component {
           {this.renderMenu()}
           {this.renderContent()}
         </div>
+        <ChromeToggleContainerStyled>
+          <FormCheckableInput
+            checked={this.props.showChrome}
+            onChange={this.onToggleChrome}
+            type="checkbox"
+          >
+            Show Example Chrome
+          </FormCheckableInput>
+        </ChromeToggleContainerStyled>
       </div>
     );
   }
 }
 
-let mapStateToProps = function({menu}) {
-  return {
-    menuItems: menu.items,
-  };
-};
-
-export default connect(mapStateToProps)(Application);
+export default connect(mapStateToProps, mapDispatchToProps)(Application);