+++ CHANGELOG.md
@@ -6,6 +6,8 @@ There are some key used within the changelog which are as following:
 
 # IN PROGRESS
 
+* added ability to hide example chrome #8
+
 ## 0.1.0
 
 * initial commit which is a very heavily modified version of [nucleus-react](https://github.com/ryanzec/nucleus-react)